p6 move attach/detach to SQLiteDatastore
